made non draggable functionality
